moved all lazybot-plus class into plus package
