+++ src/main/java/me/aloic/lazybot/entity/command/AddScorePlus.java
@@ -2,9 +2,7 @@ package me.aloic.lazybot.entity.command;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import me.aloic.lazybot.osu.dao.entity.dto.lazybot.LazybotScorePerformance;
-import me.aloic.lazybot.osu.dao.entity.vo.PPPlusPerformance;
-import me.aloic.lazybot.osu.dao.entity.vo.PlayerInfoVO;
+import me.aloic.lazybot.osu.dao.entity.dto.plus.LazybotScorePerformance;
 import me.aloic.lazybot.osu.dao.entity.vo.ScoreVO;
